main: make the output size and HDR flag atomic
g_nOutputWidth and g_nOutputHeight are written by one thread and read by
several. The writers are the backend size paths, the DRM mode set, the SDL
and Wayland configure handlers, the headless, OpenVR and deferred
defaults, and --track-app-size's host configure path, all of which run on
the thread that owns the output size. The readers are not on it: the
Wayland input thread normalises pointer and touch coordinates against the
pair, wlserver scales touch the same way, the pipewire capture thread
compares and copies it every iteration, and mangoapp publishes it. Those
reads are a data race on a plain uint32_t, which is undefined behaviour
rather than a stale value: nothing stops the compiler hoisting the load
out of the input thread's loop or reloading it twice within one
expression, and with LTO it can see across the translation units that made
it look accidentally safe.

g_nOutputRefresh has the same shape, written by the Wayland output
listeners while the vblank nudge thread divides by it. So does
g_bOutputHDREnabled, which is read cross-thread (the OpenVR backend's
input thread, for one), and which tying it to app content turns from a
write once at startup into a write every few hundred milliseconds on the
steamcompmgr thread.

Make all four atomic. The implicit conversions mean almost every use site
compiles unchanged; a few ternaries need their arms to agree on a type.
The local extern declarations in rendervulkan.cpp and
reshade_effect_manager.cpp are updated to match.

The size pair is deliberately not made jointly atomic: a reader can still
catch a new width against an old height for one event, costing at worst
one mispositioned input event or one capture frame at a mismatched size,
both corrected on the next one. A consumer that needs the pair as one
value has to carry it as one value, which no global here does. The
single-writer property still matters for the read-modify-write size
fixups.

+++ src/Backends/WaylandBackend.cpp
@@ -1137,7 +1137,7 @@ namespace gamescope
                 oApp = LayerRect( pFrameInfo->layers.get( nAppLayer ) );
         }
 
-        const app_viewport::Rect output{ 0, 0, g_nOutputWidth, g_nOutputHeight };
+        const app_viewport::Rect output{ 0, 0, uint32_t( g_nOutputWidth ), uint32_t( g_nOutputHeight ) };
         const app_viewport::Rect viewport = app_viewport::ComputeViewport( oApp, output.uWidth, output.uHeight );
         // A host-dictated resize moves the output without moving the viewport,
         // so this is settled before the unchanged-viewport early return.
